google-genai-agent/pyproject.toml
[project]
name = "google-genai-agent"
version = "0.1.0"
description = "Google GenAI Agent - Vertex AI/Gemini for integrated RAG and seamless SDK"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]


